Add RGear, UGear, and MGear rollsums, and use a weighted geometric average score.
In RESULTS.rst document RGear, UGear, and MGear, performance. Add separate
sections for hashing and chunking performance.

In cmphash.py change the score from a simple geometric average to a weighted
geometric average. Add tests for 16 byte windows and for RGear, UGear and
MGear.

In rollsum.py add UGear and MGear rollsums. In Stats change the perf property to use the
binomial instead of poisson expected variance. Add a weight property that is
roughly proportional to the number of digits accuracy for calculating weighted
scores. Add cmdline support for MGear and UGear.

Update data/cmphash.rst with newly generated results.